% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataCleanup.R
\name{accessdenied}
\alias{accessdenied}
\title{Handling of "access denied" errors when downloading files}
\usage{
accessdenied(FileID)
}
\arguments{
\item{FileID}{The DigitalFileID used in the NPS Data Store Application.}
}
\description{
\code{accessdenied} adds a "true" value to the "RestrictedFiles" table in the master scraper
database.
}
\details{
This function is used to flag a DigitalFileID when an attempted download throws an "access denied" error.
}
\examples{
accessdenied(474037)
}
