% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpeciesDictionary.R
\name{createSpeciesDictionary}
\alias{createSpeciesDictionary}
\title{Create species dictionary file}
\usage{
createSpeciesDictionary()
}
\description{
\code{dicSpecies} creates a dictionary file of species to be used for text analysis.
}
\details{
This function is used to create a LIWC-compliant dictionary file of latin binomial species
names based on taxa currently included in the NPSpecies master taxonomy table.
The dictionary file includes the two-word species names as "words" and the NPSpeciestaxonomic
group as the "sentiment" column in the LWIC file.

Note that this is a maintenance function and should only be run while working on the
pdfSpeciesScraper project to perform updates to the dicSpecies.rda file included with the
distribution.
}
\examples{
createSpeciesDictionary()
}
