% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpeciesDictionary.R
\name{getNPSpeciesTaxa}
\alias{getNPSpeciesTaxa}
\title{Get taxon list from NPSpecies database}
\usage{
getNPSpeciesTaxa()
}
\description{
\code{getNPSpeciesTaxa} pulls the latest taxonomy from the NPSpecies/Taxonomy application
}
\details{
This function is used to pull the most-recent list of taxonomy from theNPSpecies/Taxonomy
application.

Note that this is a maintenance function and should only be run while working on the
pdfSpeciesScraper project to perform updates to the NPSpeciesJustSpecies.rda or dicSpecies.rda
files included with the distribution.
}
\examples{
getNPSpeciesTaxa()
}
