% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScraperScript.R
\name{runlots}
\alias{runlots}
\title{Wrapper function to process multiple runs of the scrapeaway function.}
\usage{
runlots(chunksize, chunks)
}
\arguments{
\item{chunksize}{The number of files to download in a single call of the function}

\item{chunks}{The number of runs to do}
}
\description{
\code{runlots} processes several runs (chunks) of the scrapeaway function.
}
\details{
This function is used to run multiple runs (chunks) of the scrapeaway function. The chunksize parameter is passed
on to the scrapeaway function.
}
\examples{
runlots(50,100)
}
