% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScraperScript.R
\name{scrapeaway}
\alias{scrapeaway}
\title{Primary function to download and analyze a PDF file.}
\usage{
scrapeaway(chunksize)
}
\arguments{
\item{chunksize}{The number of files to download in a single call of the function}
}
\description{
\code{scrapeaway} processes a batch of pdf files to test for the presence of mentions of species names.
}
\details{
This function is used to download files from the NPS Data Store Application, extract text from pdf files, and then compare
those files to the a taxonomic (species-level) dictionary distributed with this package. This function is called by a sister
function (runlots) where the user can set the number of files to process in a single run. The recommended number of files is
between 25 and 100, depending on available bandwidth and processing power.
}
\examples{
scrapeaway(50)
}
