% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpeciesDictionary.R
\name{taxaCategories}
\alias{taxaCategories}
\title{Get species for a category of taxa in NPSpecies}
\usage{
taxaCategories(category)
}
\arguments{
\item{category}{is one of fifteen taxonomic categories within the NPSpecies/Taxonomy application:

\describe{
  \item{"Bird"}{Birds}
  \item{"Fish"}{Fishes}
  \item{"Insect"}{Insects}
  \item{"Other Non-vertebrates"}{Other non-vertebrate species}
  \item{"Non-vascular Plant"}{Non-vascular plants}
  \item{"Fungi"}{Fungi}
  \item{"Protozoa"}{Protozoa}
  \item{"Vascular Plant"}{Vascular plants}
  \item{"Slug/Snail"}{Slugs and snails}
  \item{"Amphibian"}{Amphibians}
  \item{"Reptile"}{Reptiles}
  \item{"Mammal"}{Mammals}
  \item{"Crab/Lobster/Shrimp"}{Crabs, lobsters, and shrimp}
  \item{"Spider/Scorpion"}{Spiders and scorpions}
  \item{"Archaea"}{Prokaryotic microorganisms}
}}
}
\description{
\code{taxaCategories} returns all taxa within a taxonomic category in NPSpecies
}
\details{
This function is used to extract latin species names from the NPSpecies master taxonomy table for
any given taxonomic group. Used by the \code{dicSpecies} function when building the dictionary
for text comparison.
}
\examples{
taxaCategories("Bird")
}
