% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpeciesDictionary.R
\name{updateSpeciesDictionary}
\alias{updateSpeciesDictionary}
\title{Update species dictionary file}
\usage{
updateSpeciesDictionary()
}
\description{
\code{dicSpecies} updates a dictionary file of species to be used for text analysis.
}
\details{
This function is used to update the LIWC-compliant dictionary file of latin binomial species
names based on taxa currently included in the NPSpecies master taxonomy table.
The dictionary file includes the two-word species names as "words" and the NPSpeciestaxonomic
group as the "sentiment" column in the LWIC file.

Note that this is a maintenance function and should only be run while working on the
pdfSpeciesScraper project to perform updates to the dicSpecies.rda file included with the
distribution.
}
\examples{
updateSpeciesDictionary()
}
